Added check_mode to Parallel.GAMIT

+++ parallel_gamit/pyParallelGamit.py
@@ -27,6 +27,7 @@ from Utils import process_date
 from Utils import process_stnlist
 from Utils import parseIntSet
 from pyStation import Station
+from pyStation import StationCollection
 from pyETM import pyETMException
 import pyArchiveStruct
 import logging
@@ -169,10 +170,10 @@ def purge_solutions(JobServer, args, dates, GamitConfig):
         JobServer.close_cluster()
 
 
-def station_list(NetworkConfig, dates):
+def station_list(stations, dates):
 
-    stations = process_stnlist(cnn, NetworkConfig['stn_list'].split(','))
-    stn_obj = []
+    stations = process_stnlist(cnn, stations)
+    stn_obj = StationCollection()
 
     # use the connection to the db to get the stations
     for Stn in tqdm(sorted(stations), ncols=80, disable=None):
@@ -180,60 +181,24 @@ def station_list(NetworkConfig, dates):
         NetworkCode = Stn['NetworkCode']
         StationCode = Stn['StationCode']
 
-        # apply date -1 and date + 1 to avoid problems with files ending at 00:00 of date + 1
         rs = cnn.query(
             'SELECT * FROM rinex_proc WHERE "NetworkCode" = \'%s\' AND "StationCode" = \'%s\' AND '
-            '"ObservationSTime" >= \'%s\' AND "ObservationETime" <= \'%s\''
-            % (NetworkCode, StationCode, (dates[0] - 1).first_epoch(), (dates[1] + 1).last_epoch()))
+            '("ObservationYear", "ObservationDOY") BETWEEN (%s) AND (%s)'
+            % (NetworkCode, StationCode,
+               dates[0].yyyy() + ', ' + dates[0].ddd(),
+               dates[1].yyyy() + ', ' + dates[1].ddd()))
 
         if rs.ntuples() > 0:
-
             tqdm.write(' -- %s.%s -> adding...' % (NetworkCode, StationCode))
             try:
                 stn_obj.append(Station(cnn, NetworkCode, StationCode, dates))
-
             except pyETMException:
                 tqdm.write('    %s.%s -> station exists, but there was a problem initializing ETM.'
                            % (NetworkCode, StationCode))
         else:
             tqdm.write(' -- %s.%s -> no data for requested time window' % (NetworkCode, StationCode))
 
-    # analyze duplicate names in the list of stations
-    stn_obj = check_station_codes(stn_obj)
-
     return stn_obj
-
-
-def check_station_codes(stn_obj):
-
-    for i, stn1 in enumerate(stn_obj[:-1]):
-
-        for stn2 in stn_obj[i+1:]:
-            if stn1.NetworkCode != stn2.NetworkCode and stn1.StationCode == stn2.StationCode:
-                # duplicate StationCode (different Network), produce Alias
-                unique = False
-                while not unique:
-                    stn1.generate_alias()
-                    # compare again to make sure this name is unique
-                    unique = compare_aliases(stn1, stn_obj)
-
-    return stn_obj
-
-
-def compare_aliases(Station, AllStations):
-
-    # make sure alias does not exists as alias and station code
-
-    for stn in AllStations:
-
-        # this if prevents comparing against myself, although the station is not added until after
-        # the call to CompareAliases. But, just in case...
-        if stn.StationCode != Station.StationCode and stn.NetworkCode != Station.NetworkCode and \
-                        Station.StationAlias == stn.StationAlias or Station.StationAlias == stn.StationCode:
-            # not unique!
-            return False
-
-    return True
 
 
 def main():
@@ -247,9 +212,21 @@ def main():
     parser.add_argument('-dp', '--date_parser', type=str, nargs=2, metavar='{year} {doys}',
                         help="Parse date using ranges and commas (e.g. 2018 1,3-6). "
                              "Cannot cross year boundaries")
-    parser.add_argument('-e', '--exclude', type=str, nargs='+', metavar='station',
+    parser.add_argument('-e', '--exclude', type=str, nargs='+', metavar='{station}',
                         help="List of stations to exclude from this processing (e.g. -e igm1 lpgs vbca)")
-    parser.add_argument('-p', '--purge', action='store_true',
+    parser.add_argument('-c', '--check_mode', type=str, nargs='+', metavar='{station}',
+                        help="Check station(s) mode. If station(s) are not present in the GAMIT polyhedron, "
+                             "(i.e. the RINEX file(s) were missing at the time of the processing) Parallel.GAMIT will "
+                             "add the station to the closest subnetwork(s) and reprocess them. If station(s) were "
+                             "present at the time of the processing but failed to process (i.e. they are in the "
+                             "missing stations list), these subnetworks will be reprocessed to try to obtain a "
+                             "solution. Station list provided in the cfg is ignored in this mode. Therefore, changes "
+                             "in the station list will not produce any changes in network configuration. Purge not "
+                             "allowed when using this mode. (Syntax: -c igm1 lpgs rms.vbca)")
+    parser.add_argument('-i', '--ignore_missing', action='store_true',
+                        help="When using check mode or processing existing sessions, ignore missing stations. In other "
+                             "words, do not try to reprocess sessions that have missing solutions.")
+    parser.add_argument('-p', '--purge', action='store_true', default=False,
                         help="Purge year doys from the database and directory structure and re-run the solution.")
     parser.add_argument('-dry', '--dry_run', action='store_true',
                         help="Generate the directory structures (locally) but do not run GAMIT. "
@@ -308,18 +285,29 @@ def main():
         Utils.print_columns(exclude)
         GamitConfig.NetworkConfig['stn_list'] += ',-' + ',-'.join(exclude)
 
+    # initialize stations in the project
+    stations = station_list(GamitConfig.NetworkConfig['stn_list'].split(','), drange)
+
+    check_station_list = args.check_mode
+    if check_station_list is not None:
+        print(' >> Check mode. List of stations to check for selected days:')
+        Utils.print_columns(check_station_list)
+        check_stations = station_list(check_station_list, drange)
+    else:
+        check_stations = None
+
     if args.dry_run is not None:
         dry_run = args.dry_run
     else:
         dry_run = False
 
-    if not dry_run:
+    if not dry_run and check_stations is None:
         # ignore if calling a dry run
         # purge solutions if requested
         purge_solutions(JobServer, args, dates, GamitConfig)
-
-    # initialize stations in the project
-    stations = station_list(GamitConfig.NetworkConfig, drange)
+    else:
+        if args.purge:
+            print(' >> Dry run or check mode activated. Cannot purge solutions in these modes.')
 
     tqdm.write(' >> %s Creating GAMIT session instances, please wait...'
                % datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
@@ -336,7 +324,7 @@ def main():
         if not os.path.exists(pwd):
             os.makedirs(pwd)
 
-        net_object = Network(cnn, archive, GamitConfig, stations, date)
+        net_object = Network(cnn, archive, GamitConfig, stations, date, check_stations, args.ignore_missing)
 
         sessions += net_object.sessions
 
@@ -349,7 +337,7 @@ def main():
     # print_summary(stations, sessions, drange)
 
     # run the job server
-    # ExecuteGamit(sessions, JobServer, dry_run)
+    ExecuteGamit(sessions, JobServer, dry_run)
 
     # execute globk on doys that had to be divided into subnets
     if not args.dry_run:
@@ -712,9 +700,9 @@ def ExecuteGamit(Sessions, JobServer, dry_run=False):
 
             JobServer.submit(task, task.params['DirName'], task.date.year, task.date.doy, dry_run)
 
-            tqdm.write(' -- %s Submitting %s %s %02i for processing'
+            tqdm.write(' -- %s Submitting %s %s %s%02i for processing'
                        % (datetime.now().strftime('%Y-%m-%d %H:%M:%S'),
-                          GamitSession.NetName, GamitSession.date.yyyyddd(),
+                          GamitSession.NetName, GamitSession.date.yyyyddd(), GamitSession.org,
                           GamitSession.subnet if GamitSession.subnet is not None else 0))
             # save it just in case...
             # gtasks.append(task)
